+++ sandbox/src/sharedmem_ipc_server.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -50,7 +50,7 @@ class SharedMemIPCServer {
 
   // Initializes the server structures, shared memory structures and
   // creates the kernels events used to signal the IPC.
-  bool Init(void* shared_mem, size_t shared_size, size_t channel_size);
+  bool Init(void* shared_mem, uint32 shared_size, uint32 channel_size);
 
  private:
   // Allow tests to be marked DISABLED_. Note that FLAKY_ and FAILS_ prefixes
@@ -78,7 +78,7 @@ class SharedMemIPCServer {
     // This channel server pong event.
     HANDLE pong_event;
     // The size of this channel.
-    size_t channel_size;
+    uint32 channel_size;
     // The pointer to the actual channel data.
     char* channel_buffer;
     // The pointer to the base of the shared memory.
